Fix comparison of diffrent type datatypes

+++ 0x1E-search_algorithms/103-exponential.c
@@ -18,14 +18,14 @@ int exponential_search(int *array, size_t size, int value)
 		return (0);
 
 	bound = 1;
-	while (bound < size && array[bound] <= value)
+	while (bound < (int) size && array[bound] <= value)
 	{
 		printf("Value checked array[%d] = [%d]\n", bound, array[bound]);
 		bound *= 2;
 	}
 
 	left = bound / 2;
-	right = (bound < size) ? bound : (int)size - 1;
+	right = (bound < (int) size) ? bound : (int)size - 1;
 	printf("Value found between indexes [%d] and [%d]\n", left, right);
 
 	return (binary_search2(array, left, right, value));
